requirements.txt
py_cord @ git+https://github.com/Pycord-Development/pycord
python-dateutil==2.8.2